refactor(platform): extract shared components to fix function line-count lint
Extract MinimalSidebarLayout, Vm0LogoLink, and GoogleOAuthNotice into
zero-directed-shared.tsx to deduplicate code between authorize and
connect pages, bringing both DirectedAuthorizeCard and
DirectedConnectCard under the 128-line eslint limit.
