+++ turbo/apps/platform/src/views/zero-page/zero-directed-authorize-page.tsx
@@ -1,12 +1,9 @@
-import type { ReactNode } from "react";
 import { useGet, useSet, useLastLoadable } from "ccstate-react";
 import {
   CONNECTOR_TYPES,
   type ConnectorType,
   isGoogleOAuthConnector,
 } from "@vm0/core";
-import { AvatarSvgPreview } from "./avatar-svg-preview.tsx";
-import { getAvatarPresets } from "./zero-avatars.ts";
 import { ConnectorIcon } from "./components/settings/connector-icons.tsx";
 import {
   allConnectorTypes$,
@@ -24,99 +21,12 @@ import {
   authorizeConnector$,
 } from "../../signals/connectors-page/directed-authorize-type.ts";
 import { pageSignal$ } from "../../signals/page-signal.ts";
-import { handleZeroAccountAction$ } from "../../signals/zero-page/zero-nav.ts";
-import {
-  orgManageDialogOpen$,
-  setOrgManageDialogOpen$,
-} from "../../signals/zero-page/settings/org-manage-dialog.ts";
-import { OrgManageDialog } from "./components/org-manage/org-manage-dialog.tsx";
-import { ZeroOrgSwitcher } from "./zero-org-switcher.tsx";
-import { AccountDropdown } from "./zero-sidebar.tsx";
-import { VM0ClerkProvider } from "../clerk/clerk-provider.tsx";
 import { IconCheck, IconLoader2 } from "@tabler/icons-react";
-import { Link } from "../router/link.tsx";
-
-// ---------------------------------------------------------------------------
-// Layout (shared with directed connect page)
-// ---------------------------------------------------------------------------
-
-function MinimalSidebarLayout({ children }: { children: ReactNode }) {
-  const onAccountAction = useSet(handleZeroAccountAction$);
-  const dialogOpen = useGet(orgManageDialogOpen$);
-  const setDialogOpen = useSet(setOrgManageDialogOpen$);
-  const pageSignal = useGet(pageSignal$);
-
-  return (
-    <div className="zero-app flex h-dvh w-full bg-background">
-      <OrgManageDialog
-        open={dialogOpen}
-        onOpenChange={(open) => {
-          detach(setDialogOpen(open, pageSignal), Reason.DomCallback);
-        }}
-      />
-      <VM0ClerkProvider>
-        <aside className="zero-nav hidden md:flex h-full w-[255px] shrink-0 flex-col bg-sidebar">
-          <div className="shrink-0 px-2 pt-1.5">
-            <ZeroOrgSwitcher />
-          </div>
-          <div className="flex-1" />
-          <div className="p-2">
-            <AccountDropdown onAccountAction={onAccountAction} />
-          </div>
-        </aside>
-      </VM0ClerkProvider>
-      <div className="flex flex-1 flex-col min-w-0 min-h-0 zero-workspace-bg">
-        {children}
-      </div>
-    </div>
-  );
-}
-
-function Vm0Logo() {
-  return (
-    <svg
-      width="82"
-      height="20"
-      viewBox="0 0 100 30"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      aria-label="VM0"
-    >
-      <path
-        d="M13.3915 0.0627979C13.2455 -0.0209506 13.0657 -0.020839 12.9198 0.0630906L1.0053 6.91543C0.692394 7.09539 0.690093 7.54442 1.00114 7.72755L12.9156 14.7423C13.0636 14.8295 13.2475 14.8296 13.3957 14.7426L25.3445 7.72785C25.6562 7.54485 25.6539 7.09497 25.3404 6.91514L13.3915 0.0627979Z"
-        fill="#ED4E01"
-      />
-      <path
-        d="M0.710495 8.33374L12.6479 15.2595C12.7944 15.3445 12.8846 15.5015 12.8846 15.6715L12.8843 29.5237C12.8843 29.8899 12.4897 30.1187 12.1741 29.9356L0.236691 23.0096C0.0902206 22.9246 -3.46036e-06 22.7676 0 22.5977L0.00028208 8.74568C0.000289537 8.37949 0.394855 8.15064 0.710495 8.33374Z"
-        fill="#ED4E01"
-      />
-      <path
-        d="M24.947 21.6772C24.947 21.9507 24.8017 22.2036 24.5655 22.3415L16.2103 27.219C15.6975 27.5184 15.0533 27.1485 15.0533 26.5547L15.0531 16.7842C15.0531 16.5107 15.1983 16.2578 15.4345 16.1199L23.7897 11.2425C24.3025 10.9431 24.9468 11.313 24.9468 11.9068L24.947 21.6772ZM13.6541 16.3426V29.5279C13.6541 29.8852 14.0308 30.1106 14.3391 29.9444L14.3538 29.9362L25.5769 23.3654C26.25 22.9808 26.3462 22.6924 26.3459 22.1188L26.3459 8.93378C26.3459 8.57084 25.9572 8.344 25.6462 8.52548L14.4231 15.0001C14.0385 15.2885 13.6539 15.577 13.6541 16.3426Z"
-        fill="#ED4E01"
-      />
-      <path
-        d="M25.9616 10.58L15.2113 28.4616L14.2308 27.8817L24.981 10.0001L25.9616 10.58Z"
-        fill="#ED4E01"
-      />
-      <path
-        d="M42.1865 25L34.3459 5H37.4651L43.7887 21.4575L50.1264 5H53.2315L45.3908 25H42.1865Z"
-        fill="currentColor"
-      />
-      <path
-        d="M66.9877 25L59.4023 10.3417V25H56.4957V5H59.6716L67.413 20.0628L75.1686 5H78.3304V25H75.438V10.3417L67.8526 25H66.9877Z"
-        fill="currentColor"
-      />
-      <path
-        d="M99.3459 22.1409C99.3459 22.5314 99.2703 22.9033 99.1191 23.2566C98.9678 23.6007 98.7599 23.9028 98.4952 24.1632C98.2305 24.4235 97.9186 24.6281 97.5594 24.7768C97.2097 24.9256 96.8363 25 96.4393 25H86.2735C85.8765 25 85.4984 24.9256 85.1392 24.7768C84.7894 24.6281 84.4822 24.4235 84.2176 24.1632C83.9529 23.9028 83.745 23.6007 83.5937 23.2566C83.4425 22.9033 83.3669 22.5314 83.3669 22.1409V7.85914C83.3669 7.46862 83.4425 7.10135 83.5937 6.75732C83.745 6.404 83.9529 6.10181 84.2176 5.85077C84.4822 5.59042 84.7894 5.38587 85.1392 5.2371C85.4984 5.07903 85.8765 5 86.2735 5H96.4393C96.8363 5 97.2097 5.07903 97.5594 5.2371C97.9186 5.38587 98.2305 5.59042 98.4952 5.85077C98.7599 6.10181 98.9678 6.404 99.1191 6.75732C99.2703 7.10135 99.3459 7.46862 99.3459 7.85914V22.1409ZM86.2735 7.85914V22.1409H96.4393V7.85914H86.2735Z"
-        fill="currentColor"
-      />
-      <path
-        d="M94.8994 6.79107L97.1494 8.06891L87.8973 23.8325L85.6473 22.5547L94.8994 6.79107Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-}
+import {
+  MinimalSidebarLayout,
+  Vm0LogoLink,
+  GoogleOAuthNotice,
+} from "./zero-directed-shared.tsx";
 
 // ---------------------------------------------------------------------------
 // Main card
@@ -164,10 +74,8 @@ function DirectedAuthorizeCard() {
 
   const handleAuthorize = () => {
     if (isConnected) {
-      // Connector already connected — just authorize for this agent
       detach(authorize(connectorType, agentId, signal), Reason.DomCallback);
     } else {
-      // Need to connect first, then authorize
       detach(
         (async () => {
           await connect(connectorType, signal);
@@ -181,9 +89,7 @@ function DirectedAuthorizeCard() {
   return (
     <div className="fixed inset-0 z-10 flex items-center justify-center pointer-events-none">
       <div className="pointer-events-auto flex w-[430px] max-w-[calc(100%-48px)] flex-col items-center gap-12 rounded-[20px] border border-border bg-background px-6 py-12 text-center">
-        <Link pathname="/connectors" className="no-underline text-foreground">
-          <Vm0Logo />
-        </Link>
+        <Vm0LogoLink />
         <div className="flex w-full flex-col gap-4">
           <div className="flex flex-col items-center gap-2.5">
             {isLoading ? (
@@ -207,32 +113,7 @@ function DirectedAuthorizeCard() {
                 {!isAuthorized &&
                   !isConnected &&
                   isGoogleOAuthConnector(connectorType) && (
-                    <div className="w-full mt-2 flex flex-col gap-3 rounded-lg bg-muted/50 px-4 py-4 text-left">
-                      <AvatarSvgPreview
-                        config={getAvatarPresets()[0]}
-                        size={36}
-                        className="h-9 w-9 rounded-full"
-                        alt="Zero"
-                      />
-                      <div className="text-xs leading-relaxed text-muted-foreground space-y-1.5">
-                        <p>
-                          Google will show a security warning because
-                          we&rsquo;re still completing their app verification.
-                          This is a standard review process for new integrations
-                          and does not affect your security.
-                        </p>
-                        <p>
-                          To continue, select{" "}
-                          <strong className="text-foreground">Advanced</strong>{" "}
-                          then{" "}
-                          <strong className="text-foreground">
-                            Go to vm0.ai (unsafe)
-                          </strong>
-                          . We only request the minimum permissions needed, and
-                          your credentials are encrypted at rest.
-                        </p>
-                      </div>
-                    </div>
+                    <GoogleOAuthNotice />
                   )}
               </>
             )}
